# infra/terraform/modules/apim/openapi/gateway.yaml
openapi: 3.0.0
info:
  title: Carpeta Ciudadana Gateway API
  description: API Gateway for Carpeta Ciudadana operator
  version: 1.0.0
  contact:
    name: Carpeta Ciudadana Team
    email: support@carpeta-ciudadana.com
servers:
  - url: https://{apim-name}.azure-api.net/api
    description: API Management Gateway
paths:
...
  /documents:
    post:
      summary: Upload document
      description: Get pre-signed URL for document upload
      operationId: upload-document
      tags:
        - Documents
      x-azure-backend:
        serviceUrl: "{{metadata-backend}}"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '200':
          description: Pre-signed URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /documents/{document_id}:
    get:
      summary: Download document
      description: Get pre-signed URL for document download
      operationId: download-document
      tags:
        - Documents
      x-azure-backend:
        serviceUrl: "{{metadata-backend}}"
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pre-signed URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDownloadResponse'
        '404':
          description: Document not found
...
components:
  schemas:
...
    DocumentUploadRequest:
      type: object
      required:
        - citizen_id
        - file_name
        - content_type
      properties:
        citizen_id:
          type: integer
        file_name:
          type: string
        content_type:
          type: string
        metadata:
          type: object
    DocumentUploadResponse:
      type: object
      properties:
        upload_url:
          type: string
        document_id:
          type: string
        expires_at:
          type: string
          format: date-time
    DocumentDownloadResponse:
      type: object
      properties:
        download_url:
          type: string
        expires_at:
          type: string
          format: date-time
...